feat: 카테고리 별 최저가격 브랜드와 상품 가격, 총액을 조회 구현

+++ src/main/java/org/lsh/brandproduct/domain/ProductModel.java
@@ -1,4 +1,4 @@
 package org.lsh.brandproduct.domain;
 
-public record ProductModel(String brandName, ProductCategory category, double price, int quantity) {
+public record ProductModel(String brandName, ProductCategory category, Integer price, int quantity) {
 }
